# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the api documentation of task manager project.",
        "title": "Task Manager API Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "description": "Get all tasks of user based on query parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get all tasks",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "return tasks created by you if createdByMe set to true otherwise false.",
                        "name": "createdByMe",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of tasks to return per page (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term to filter tasks",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter tasks by status (TO-DO, In-PROGRESS, COMPLETED, CLOSED)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Sort tasks by create time (true for ascending, false for descending)",
                        "name": "sortByFilter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks fetched successfully.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateTask API is made for creating a new task in the task manager application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Create New Task",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title of the task (min length: 4, max length: 48)",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Description of the task (min length: 12, max length: 196)",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the individual assignee",
                        "name": "assigneeIndividual",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "ID of the team assignee",
                        "name": "assigneeTeam",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Status of the task (TO-DO, In-PROGRESS, COMPLETED, CLOSED)",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Priority of the task (LOW, MEDIUM, HIGH, VERY HIGH)",
                        "name": "priority",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task created successfully.",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request, either data is not valid or assignee privacy is Private.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/": {
            "put": {
                "description": "Update a task based on provided parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Update a task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "TaskID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title of the task (min length: 4, max length: 48)",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Description of the task (min length: 12, max length: 196)",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "ID of the individual assignee",
                        "name": "assigneeIndividual",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "ID of the team assignee",
                        "name": "assigneeTeam",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Status of the task (TO-DO, In-PROGRESS, COMPLETED, CLOSED)",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Priority of the task (LOW, MEDIUM, HIGH, VERY HIGH)",
                        "name": "priority",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "403": {
                        "description": "Not allowed to update task",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/team/{TeamID}": {
            "get": {
                "description": "Get all tasks of a team based on query parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get all tasks of a team",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003caccess_token\u003e",
                        "description": "Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "TeamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of tasks to return per page (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term to filter tasks",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter tasks by status (TO-DO, In-PROGRESS, COMPLETED, CLOSED)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Sort tasks by create time (true for ascending, false for descending)",
                        "name": "sortByFilter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks fetched successfully.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errorhandling.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "Bad request"
                },
                "error": {
                    "type": "string",
                    "example": "Corresponding Error Message will Show Here"
                }
            }
        },
        "response.SuccessResponse": {
            "description": "Send success response to client with corresponding message and id(if any).",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "200 OK"
                },
                "id": {
                    "type": "integer",
                    "example": 974751326021189896
                },
                "message": {
                    "type": "string",
                    "example": "Task Created Successfully."
                }
            }
        },
        "response.Task": {
            "description": "Task information with title, description, deadline, assignee, status, priority.",
            "type": "object",
            "properties": {
                "assigneeIndividual": {
                    "type": "integer",
                    "example": 974751326021189123
                },
                "assigneeTeam": {
                    "type": "integer",
                    "example": 974751326021189234
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-03-25T22:59:59.000Z"
                },
                "createdBy": {
                    "type": "integer",
                    "example": 974751326021189896
                },
                "deadline": {
                    "type": "string",
                    "example": "2024-03-25T22:59:59.000Z"
                },
                "description": {
                    "type": "string",
                    "example": "Create Task Manager Project with GoLang as Backend."
                },
                "id": {
                    "type": "integer",
                    "example": 974751326021189496
                },
                "priority": {
                    "type": "string",
                    "example": "High"
                },
                "status": {
                    "type": "string",
                    "example": "TO-DO"
                },
                "title": {
                    "type": "string",
                    "example": "GoLang project: Task Manager"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2024-03-26T12:49:539.000Z"
                },
                "updatedBy": {
                    "type": "integer",
                    "example": 974751326021189896
                }
            }
        },
...
    }
}